Removed templating, added templated source code within each example
