+++ packages/terra-date-picker/src/terra-dev-site/doc/example/DatePickerExcludeDates.jsx
@@ -1,12 +1,52 @@
 import React from 'react';
 import moment from 'moment';
-import DatePickerExampleTemplate from '../common/DatePickerExampleTemplate';
+import PropTypes from 'prop-types';
+// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
+import DatePicker from 'terra-date-picker/lib/DatePicker';
 
-const DatePickerExample = () => (
-  <DatePickerExampleTemplate
+const propTypes = {
+  /**
+   * The current DatePicker date if selected. Use for the selected date message.
+   */
+  selectedDate: PropTypes.node,
+};
+
+const defaultProps = {
+  selectedDate: '',
+};
+
+class DatePickerExample extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { date: this.props.selectedDate };
+    this.handleDateChange = this.handleDateChange.bind(this);
+  }
+
+  handleDateChange(event, date) {
+    this.setState({ date });
+  }
+
+  render() {
+    return (
+      <div>
+        <p>Selected ISO Date: <span style={{ display: 'inline-block' }}>{this.state.date}</span></p>
+        <DatePicker
+          name="date-input"
+          onChange={this.handleDateChange}
+          {...this.props}
+        />
+      </div>
+    );
+  }
+}
+
+DatePickerExample.propTypes = propTypes;
+DatePickerExample.defualtProps = defaultProps;
+
+const DatePickerExampleExcludeDates = () => (
+  <DatePickerExample
     excludeDates={[moment().subtract(1, 'days').format(), moment().add(1, 'days').format()]}
   />
 );
 
-
-export default DatePickerExample;
+export default DatePickerExampleExcludeDates;
